Refactor Pinia stores
