+++ src/App.vue
@@ -16,7 +16,7 @@
 
         <v-list-item title="Experiments" to="/experiments" link />
         <v-list-item title="Final results" link />
-        <v-list-item title="Groups" link />
+        <v-list-item title="Groups" to="/groups" link />
         <v-list-item title="Paper Submission" link />
         <v-list-item title="Papers" link />
       </v-list>
