+++ src/main.ts
@@ -15,6 +15,7 @@ import Overview from "@/components/AttendeeOverview.vue";
 import ExercisesOverview from "@/components/ExercisesOverview.vue";
 import ExperimentsOverview from "@/components/ExperimentsOverview.vue";
 import DepartmentOverview from "@/components/DepartmentOverview.vue";
+import GroupsOverview from "@/components/GroupsOverview.vue";
 import IndividualFinalResult from "@/components/IndividualFinalResult.vue";
 
 const vuetify = createVuetify({
@@ -58,6 +59,11 @@ const routes = [
     path: "/experiments",
     name: "Experiments",
     component: ExperimentsOverview,
+  },
+  {
+    path: "/groups",
+    name: "Groups",
+    component: GroupsOverview,
   },
   {
     path: "/attendee/:id",
